blueprint: don't undelete deleted/unlocked items on sync
 - have files, quizzes, learning outcomes, outcome groups,
   and modules in a blueprint course. no items should be
   locked.
 - ensure outcome group changes appear in the unsynced
   changes list correctly
 - sync to an associated course
 - delete items in the associated course
 - change the same items in the blueprint in some way
   (i.e., rename them)
 - re-sync
 - the deleted items should not be undeleted in the
   associated course
 - non-deleted outcome groups should not gain a link
   to a deleted outcome (that shows only in the left pane)

fixes ADMIN-559

Reviewed-on: https://gerrit.instructure.com/133374
Tested-by: Jenkins
QA-Review: Deepeeca Soundarrajan <dsoundarrajan@instructure.com>
Product-Review: Jeremy Stanley <jeremy@instructure.com>

+++ app/models/importers/quiz_importer.rb
@@ -166,8 +166,10 @@ module Importers
       # there might not be an import id if it's just a text-only type...
       item ||= Quizzes::Quiz.where(context_type: context.class.to_s, context_id: context, id: hash[:id]).first if hash[:id]
       item ||= Quizzes::Quiz.where(context_type: context.class.to_s, context_id: context, migration_id: hash[:migration_id]).first if hash[:migration_id]
-      if item && !allow_update
-        if item.deleted?
+      item ||= context.quizzes.temp_record
+      item.mark_as_importing!(migration)
+      if item
+        if !allow_update && item.deleted?
           item.workflow_state = (hash[:available] || !item.can_unpublish?) ? 'available' : 'unpublished'
           item.saved_by = :migration
           item.quiz_groups.destroy_all
@@ -175,8 +177,6 @@ module Importers
           item.save
         end
       end
-      item ||= context.quizzes.temp_record
-      item.mark_as_importing!(migration)
       new_record = item.new_record? || item.deleted?
 
       hash[:due_at] ||= hash[:due_date] if hash.has_key?(:due_date)
